Remove include:: replacements hack (#2875)
We currently need to replace the include:: directives in the cpython
documentation source files to fix an issue that sphinx/docutils have
with relative source files after changing the app.srcdir in our conf.py.

Instead of doing this, let's define the app.srcdir property as an
absolute path. This removes the issue altogether, yielding a correct
build without the need to modify the documentation source files.

This change seems to have the nice side effect that successive builds
are not incremental rather than starting from scratch.

Edit: additionally, we now define sphinx's `app.srcdir` as a Path, which
seems to be a requirement at least in 8.1.3, which is what we''l use for
building the 3.13 docs. See
https://github.com/sphinx-doc/sphinx/issues/13018.

Closes #1844

---------

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>